Refactor : Winning 결과 dto 로 변경
